process_data.py and train_classifier.py updated
The process_data.py changed the merger to be inner once the categories are expanded since otherwise get 138 null rows that will have to eliminate. For the train_classifier.py module, a los off things have been added mainly worked on the evaluation metric. The WIP is on defining the param grid for the GridSearch method before testing running the modules and the performance of the algorithm